+++ crates/parol/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Fix [#596](https://github.com/jsinger67/parol/issues/596)
 * Fix parol (par) file generation when `%t_type` is used.
 * Fix parol (par) file generation when `%nt_type` is used.
+* Fix Token alias handling in ParolGrammar
 
 ## 3.0.0 - 2025-03-01